Changes for consistent quickstart
Added random-route to all manifests and name where missing. For the cloud.gov
quickstart, I want all of these to start up with `cf push` and random-routes
so users don't get route conflicts.

Clarified in README that this is primarily a cloud.gov project.

I want to get rid of Java but I assume folks will come
here looking for that so not ready to pull plug yet.